+++ src/Forms/index.js
@@ -1,0 +1,8 @@
+import './Form.scss';
+
+import Form from './Form';
+import Label from './Label';
+
+Form.Label = Label;
+
+export default Form;