+++ src/Inputs/index.js
@@ -1,0 +1,9 @@
+import './Inputs.scss';
+
+export Text from './Text';
+export Textarea from './Textarea';
+export Password from './Password';
+export Number from './Number';
+export Select from './Select';
+export Radio from './Radio';
+export CheckBox from './CheckBox';