+++ src/index.js
@@ -2,8 +2,25 @@ import './reset.scss';
 import './colors.scss';
 import './typography.scss';
 import './helpers.scss';
+import './shadows.scss';
 
 export Alert from './Alert';
-export Button from './Buttons';
-export Card from './Cards';
+export Button from './Button';
+export Card from './Card';
+export Col from './Col';
+export Container from './Container';
+export Dropdown from './Dropdown';
+export Footer from './Footer';
+export {
+  Text,
+  Password,
+  Textarea,
+  Number,
+  CheckBox,
+  Select,
+  Radio
+} from './Inputs';
+export Nav from './Nav';
 export Preloader from './Preloader';
+export Row from './Row';
+export Table from './Table';