+++ CHANGELOG.md
@@ -1,1 +1,51 @@
 # Monument UI Changelog
+
+## Next
+#### Monument
+- adaptive monument color
+- different directions
+
+#### Inputs
+- add visibility switch for Password input component
+
+#### Buttons
+- add Button value option (remove children)
+
+#### New components
+- ButtonGroup
+- Nav
+- Footer
+- Form.Label
+- Dropdown
+
+## 0.1.0 (Soon)
+#### New Feature
+##### Buttons
+- 5 sizes: `xl`, `lg`, `md`, `sm`, `xs`
+
+##### Inputs
+- Text
+- Textarea
+- Password
+- Number
+- CheckBox
+- Radio
+- Select
+
+##### Forms
+
+##### Cards
+- Card container
+- Card title
+- Card content
+- Card image
+- Card footer
+
+##### Preloader
+- 5 sizes: `xl`, `lg`, `md`, `sm`, `xs`
+
+##### Row & Col
+
+##### Lists
+
+##### Table